requirements.txt
pandas>=1.3.0
numpy>=1.20.0
python-dateutil>=2.8.0
git+https://github.com/joshcarty/google-searchconsole
